Get a free form report

Organizes the arguments into a json string and then structures the data after the internal function makes the api call. Up to 7 dimensions at this time.

aa_freeform_report(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  date_range = c("2020-08-01", "2020-09-25"),
  dimensions = c("page", "lasttouchchannel", "mobiledevicetype"),
  metrics = c("visits", "visitors"),
  top = c(5),
  filterType = "breakdown",
  segmentId = NA,
  metricSort = "desc",
  include_unspecified = TRUE,
  debug = FALSE,
  search = NA
)

Arguments

company_id
Company Id. Taken from the global environment by default if not provided.

rsid
Adobe report suite id number. Taken from the global environment by default if not provided.

date_range
A two length vector of start and end Date objects

dimensions
Dimension to send

metrics
Metric to send

top
How many rows. Default is set to 5. If using 'daterangeday' as the first variable you can either use only one limit number or you can add a 0 as the first in the list of numbers. The function will then calculate how many days are included for you. This only works if daterangeday is the first dimension listed.

filterType
Default is 'breakdown'. This will only change if a segment is used.

segmentId
use segments to globally filter the results. Use 1 or many.

metricSort
Presorts the table by metrics. Values are either 'asc' or 'desc'.

include_unspecified
TRUE is equal to "return-nones" and is set as the default

debug
default is FALSE but set to TRUE to see the json request being sent to the Adobe API

search
The structure of a search string is specific and can be compound using "and"...more to come

Value

Data Frame