Get list of calculated metrics

Retrieve All Calculated Metrics

aa_get_calculatedmetrics(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsids = NA,
  ownerId = NA,
  filterByIds = NA,
  toBeUsedInRsid = NA,
  locale = "en_US",
  name = NA,
  tagnames = NA,
  favorite = 0,
  approved = TRUE,
  limit = 100,
  page = 0,
  sortDirection = "DESC",
  sortProperty = NA,
  expansion = NA,
  includeType = "all"
)

Arguments

company_id
Company Id. Taken from the global environment by default if not provided.

rsids
Filter list to only include calculated metrics tied to specified RSID list (comma-delimited)

ownerId
Filter list to only include calculated metrics owned by the specified loginId

filterByIds
Filter list to only include calculated metrics in the specified list (comma-delimited list of IDs) (this is the same as calculatedMetricFilter, and is overwritten by calculatedMetricFilter

toBeUsedInRsid
The report suite where the calculated metric intended to be used. This report suite will be used to determine things like compatibility and permissions. If it is not specified then the permissions will be calculated based on the union of all metrics authorized in all groups the user belongs to. If the compatibility expansion is specified and toBeUsedInRsid is not then the compatibility returned is based off the compatibility from the last time the calculated metric was saved.

locale
Locale (en_US default)

name
Filter list to only include calculated metrics that contains the Name

tagnames
Filter list to only include calculated metrics that contains one of the tags

favorite
Filter list to only include calculated metrics that are favorites (boolean)

approved
Filter list to only include calculated metrics that are approved (boolean)

limit
Number of results per page

page
Page number (base 0 - first page is "0")

sortDirection
Sort direction (ASC or DESC)

sortProperty
Property to sort by (name, modified_date, id is currently allowed)

expansion
Comma-delimited list of additional calculated metric metadata fields to include on response. (options = reportSuiteName, ownerFullName, modified, tags, definition, compatability, categories)

includeType
Include additional calculated metrics not owned by user. The "all" option takes precedence over "shared" (options = all, shared, templates)