Get list of dimensions

This will generate an extensive list of all the dimensions in the reportsuite.

aa_get_dimensions(
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "false",
  reportable = "false",
  classifiable = "false",
  expansion = "tags,allowedForReporting,categories",
  company_id = Sys.getenv("AA_COMPANY_ID")
)

Arguments

rsid
Adobe report number

locale
language - default'en_US'

segmentable
- boolean - default TRUE

reportable
- boolean - default TRUE

classifiable
- boolean - default FALSE

expansion
options c('tags', 'allowedForReporting', 'categories'))

company_id
this can be either manually defined here or utilize the global definitions