Get a single calculated metric.

Retrieve one calculated metric

aw_get_calculatedmetric_byid(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  id = NA,
  locale = "en_US",
  expansion = NA
)

Arguments

company_id
Company Id. Taken from the global environment by default if not provided.

id
The calculated metric ID to retrieve - required

locale
Locale. Default = en_US

expansion
Comma-delimited list of additional calculated metric metadata fields to include on response. (reportSuiteName, ownerFullName, modified, tags, definition, categories)