Get list of report suites

Retrieve a list of report suites and meta data about each one.

Usage

aw_get_reportsuites(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  rsids = NULL,
  rsidContains = NULL,
  limit = 10,
  page = 0,
  expansion = NULL,
  debug = FALSE
)

Arguments

company_id

Company ID. If an environment variable called AW_COMPANY_ID exists in .Renviron or elsewhere and no company_id argument is provided, then the AW_COMPANY_ID value will be used. Use get_me to get a list of available company_id values.

rsids

Filter the results to include one or more specific report suites. Specify multiple RSIDs as a vector (i.e., "rsids = c("rsid_1", rsid_2",...rsid_n")").

rsidContains

Filter the results list to only include suites that contain the specified string within the RSID. This is case-insensitive and is a simple, single string match.

limit

The number of results to return per page. This argument works in conjunction with the page argument. The default is 10.

page

The "page" of results to display. This works in conjunction with the limit argument and is zero-based. For instance, if limit = 20 and page = 1, the results returned would be 21 through 40.

expansion

Additional segment metadata fields to include in the results: name, parentRsid, currency, calendarType, timezoneZoneinfo. This argument takes a single value (e.g., expansion = "name") or a vector of values (e.g., expansion = c("name", "currency")).

debug

Include the output and input of the api call in the console for debugging. Default is FALSE

Value

A data frame of report suites and their meta data.