Validate the definition of a Calculated Metric

This function checks if a calculated metric JSON string is valid

Usage

cm_val(
  definition = NULL,
  locale = "en_US",
  migrating = FALSE,
  debug = FALSE,
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)

Arguments

definition

json string definition of a calculated metric

locale

The location of the language. en-US is default.

migrating

Include migration functions in validation. FALSE is default.

debug

This enables the api call information to show in the console for help with debugging issues. default is FALSE

rsid

Adobe report suite ID (RSID). If an environment variable called AW_REPORTSUITE_ID exists in .Renviron or elsewhere and no rsid argument is provided, then the AW_REPORTSUITE_ID value will be used. Use aw_get_reportsuites() to get a list of available rsid values.

company_id

Company ID. If an environment variable called AW_COMPANY_ID exists in .Renviron or elsewhere and no company_id argument is provided, then the AW_COMPANY_ID value will be used. Use get_me() to get a list of available company_id values.

Value

A string confirming the calculated metric is valid or is not valid.

Details

See more information here