Get Company Ids

This function will quickly pull the list of company ids that you have access to.

get_me(
  req_path = "discovery/me",
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET")
)

Arguments

req_path
The endpoint for that particular report

client_id
Set in environment args, or pass directly here

client_secret
Set in environment args, or pass directly here

Value

A data frame of company ids and company names

Examples

if (FALSE) {
get_me()
}