Get a list of user usage

This function returns the usage and access logs for a given date range within a 3 month period. The user must have Admin Console / Logs permissions (must be able to view the Usage & Access Log data in the web interface) in order to use this function.

Usage

get_usage_logs(
  startDate = Sys.Date() - 91,
  endDate = Sys.Date() - 1,
  login = NULL,
  ip = NULL,
  rsid = NULL,
  eventType = NULL,
  event = NULL,
  limit = 100,
  page = 0,
  debug = FALSE,
  company_id = Sys.getenv("AW_COMPANY_ID")
)

Arguments

startDate

Start date for the maximum of a 3 month period.

endDate

End date for the maximum of a 3 month period.

login

The login value of the user you want to filter logs by.

ip

The IP address you want to filter logs by.

rsid

The report suite ID you want to filter logs by.

eventType

The numeric id for the event type you want to filter logs by. Leaving this blank returns all events. See the Usage Logs API Guide for a complete list of event types.

event

The event description you want to filter logs by. No wildcards are permitted.

limit

The number of results to return per page. This argument works in conjunction with the page argument. The default is 10.

page

The "page" of results to display. This works in conjunction with the limit argument and is zero-based. For instance, if limit = 20 and page = 1, the results returned would be 21 through 40.

debug

Include the output and input of the api call in the console for debugging. Default is FALSE

company_id

Company ID. If an environment variable called AW_COMPANY_ID exists in .Renviron or elsewhere and no company_id argument is provided, then the AW_COMPANY_ID value will be used. Use get_me to get a list of available company_id values.

Value

A data frame of logged events and the event meta data.

Examples

if (FALSE) { # \dontrun{
get_usage_logs(startDate = Sys.Date()-91, endDate = Sys.Date()-1, limit = 100, page = 0)
} # }