Generate an Access Token for the Adobe Analytics v2.0 API

Note: aw_auth() is the primary function used for authorization. auth_oauth() and auth_jwt() should typically not be called directly.

aw_auth(type = aw_auth_with(), ...)

auth_jwt(
  file = Sys.getenv("AW_AUTH_FILE"),
  private_key = Sys.getenv("AW_PRIVATE_KEY"),
  jwt_token = NULL,
  ...
)

auth_oauth(
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET"),
  use_oob = TRUE
)

Arguments

type

Either 'jwt' or 'oauth'. This can be set explicitly, but a best practice is to run aw_auth_with() to set the authorization type as an environment variable before running aw_auth()

...

Additional arguments passed to auth functions.

file

A JSON file containing service account credentials required for JWT authentication. This file can be downloaded directly from the Adobe Console, and should minimally have the fields API_KEY, CLIENT_SECRET, ORG_ID, and TECHNICAL_ACCOUNT_ID.

private_key

Filename of the private key for JWT authentication.

jwt_token

(Optional) A custom, encoded, signed JWT claim. If used, client_id and client_secret are still required.

client_id

The client ID, defined by a global variable or manually defined

client_secret

The client secret, defined by a global variable or manually defined

use_oob

if FALSE, use a local webserver for the OAuth dance. Otherwise, provide a URL to the user and prompt for a validation code. Defaults to the value of the httr_oob_default default, or TRUE if httpuv is not installed.

Value

The path of the cached token. This is returned invisibly.

Functions

auth_jwt: Authenticate with JWT token

auth_oauth: Authorize via OAuth 2.0

See also