Get list of segments

Retrieve all segments

aw_get_segments(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  rsids = NA,
  segmentFilter = NA,
  locale = "en_US",
  name = NA,
  tagNames = NA,
  filterByPublishedSegments = "all",
  limit = 10,
  page = 0,
  sortDirection = "ASC",
  sortProperty = "id",
  expansion = NA,
  includeType = "all",
  debug = FALSE
)

Arguments

company_id
Company ID. If an environment variable called AW_COMPANY_ID exists in .Renviron or elsewhere and no company_id argument is provided, then the AW_COMPANY_ID value will be used. Use get_me to get a list of available company_id values.

rsids
Filter the list to only include segments tied to a specified RSID or list of RSIDs. Specify multiple RSIDs as a vector (i.e., "rsids = c("rsid_1", rsid_2",...rsid_n")"). Use aw_get_reportsuites to get a list of available rsid values.

segmentFilter
Filter list to only include suites in this list of segment IDs (comma-delimited)

locale
The locale that segment details should be returned in. The default is en_US.

name
Filter the list to only include segments that contain the specified name. This is case-insensitive and is a simple, single string match.

tagNames
Filter the list to only include segments that contain one of the tags.

filterByPublishedSegments
Filter the list to only include segments where the published field is set to one of the allowable values: all (the default), TRUE, or FALSE.

limit
The number of results to return per page. This argument works in conjunction with the page argument. The default is 10.

page
The "page" of results to display. This works in conjunction with the limit argument and is zero-based. For instance, if limit = 20 and page = 1, the results returned would be 21 through 40.

sortDirection
The sort direction for the results: ASC (default) for ascending or DESC for descending. (This is case insensitive, so asc and desc work as well.)

sortProperty
The property to sort the results by. Currently available values are id (default), name, and modified_date. Note that setting expansion = modified returns results with a column added called modified, which is the last date the calculated metric was modified. When using this value for sortProperty, though, the name of the argument is modified_date, because why would we expect locked-in consistency from Adobe?

expansion
Additional segment metadata fields to include in the results: reportSuiteName, ownerFullName, modified, tags, compatibility, definition, publishingStatus, definitionLastModified, and categories. This argument takes a single value (e.g., expansion = "modified") or a vector of values (e.g., expansion = c("modified", "ownerFullName")).

includeType
Include additional segments not owned by the user. Available values are all (default), shared, and templates. The all option takes precedence over "shared".

debug
Include the output and input of the api call in the console for debugging. Default is FALSE

Value

A data frame of segments and their meta data.